+++ system/mem_alloc.cpp
@@ -51,17 +51,17 @@ Arena::alloc() {
 
 void
 Arena::free(void * ptr) {
-	FreeBlock * block = (FreeBlock *)(ptr - sizeof(FreeBlock));
+	FreeBlock * block = (FreeBlock *)((UInt64)ptr - sizeof(FreeBlock));
 	block->next = _head;
 	_head = block;
 }
 
 void mem_alloc::init_thread_arena() {
-	int buf_cnt = g_thread_cnt;
+	UInt32 buf_cnt = g_thread_cnt;
 	if (buf_cnt < g_init_parallelism)
 		buf_cnt = g_init_parallelism;
 	_arenas = new Arena * [buf_cnt];
-	for (int i = 0; i < buf_cnt; i++) {
+	for (UInt32 i = 0; i < buf_cnt; i++) {
 		_arenas[i] = new Arena[SizeNum];
 		for (int n = 0; n < SizeNum; n++) {
 			assert(sizeof(Arena) == 128);
@@ -115,7 +115,7 @@ mem_alloc::get_arena_id() {
 }
 
 int 
-mem_alloc::get_size_id(int size) {
+mem_alloc::get_size_id(UInt32 size) {
 	for (int i = 0; i < SizeNum; i++) {
 		if (size <= BlockSizes[i]) 
 			return i;
@@ -129,7 +129,7 @@ void mem_alloc::free(void * ptr, uint64_t size) {
 	if (NO_FREE) {} 
 	else if (THREAD_ALLOC) {
 		int arena_id = get_arena_id();
-		FreeBlock * block = (FreeBlock *)(ptr - sizeof(FreeBlock));
+		FreeBlock * block = (FreeBlock *)((UInt64)ptr - sizeof(FreeBlock));
 		int size = block->size;
 		int size_id = get_size_id(size);
 		_arenas[arena_id][size_id].free(ptr);
